requirements.dev.txt
flake8>=3.9,<3.10
psycopg2>=2.8.6,<2.9
django-filter<22,>=2.2